+++ packages/api/src/protobuf/cerbos/cloud/epdp/v2/epdp_pb.ts
@@ -57,7 +57,7 @@ import type { Message } from "@bufbuild/protobuf";
 export const file_cerbos_cloud_epdp_v2_epdp: GenFile =
   /*@__PURE__*/
   fileDesc(
-    "Ch9jZXJib3MvY2xvdWQvZXBkcC92Mi9lcGRwLnByb3RvEhRjZXJib3MuY2xvdWQuZXBkcC52MiKqBAoGQ29uZmlnEjkKCWV2YWx1YXRvchgBIAEoCzImLmNlcmJvcy5jbG91ZC5lcGRwLnYyLkNvbmZpZy5FdmFsdWF0b3ISMwoGc2NoZW1hGAIgASgLMiMuY2VyYm9zLmNsb3VkLmVwZHAudjIuQ29uZmlnLlNjaGVtYRruAQoJRXZhbHVhdG9yEkQKB2dsb2JhbHMYASADKAsyMy5jZXJib3MuY2xvdWQuZXBkcC52Mi5Db25maWcuRXZhbHVhdG9yLkdsb2JhbHNFbnRyeRIeChZkZWZhdWx0X3BvbGljeV92ZXJzaW9uGAIgASgJEhwKFGxlbmllbnRfc2NvcGVfc2VhcmNoGAMgASgIEhUKDWRlZmF1bHRfc2NvcGUYBCABKAkaRgoMR2xvYmFsc0VudHJ5EgsKA2tleRgBIAEoCRIlCgV2YWx1ZRgCIAEoCzIWLmdvb2dsZS5wcm90b2J1Zi5WYWx1ZToCOAEavgEKBlNjaGVtYRJECgtlbmZvcmNlbWVudBgBIAEoDjIvLmNlcmJvcy5jbG91ZC5lcGRwLnYyLkNvbmZpZy5TY2hlbWEuRW5mb3JjZW1lbnQibgoLRW5mb3JjZW1lbnQSGwoXRU5GT1JDRU1FTlRfVU5TUEVDSUZJRUQQABIUChBFTkZPUkNFTUVOVF9OT05FEAESFAoQRU5GT1JDRU1FTlRfV0FSThACEhYKEkVORk9SQ0VNRU5UX1JFSkVDVBADIkQKBUVycm9yEioKBGNvZGUYASABKA4yEC5nb29nbGUucnBjLkNvZGVCCrpIB4IBBBABIAASDwoHbWVzc2FnZRgCIAEoCSKoAQoITWV0YWRhdGESHwoOY2VyYm9zX3ZlcnNpb24YASABKAlCB7pIBHICEAESJAoSY2VyYm9zX2NvbW1pdF9oYXNoGAIgASgJQgi6SAVyA5gBKBIfCg13YXNtX2NoZWNrc3VtGAMgASgJQgi6SAVyA5gBQBI0CghidWlsdF9hdBgEIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXBCBrpIA8gBASKYAQoWQ2hlY2tSZXNvdXJjZXNSZXNwb25zZRJECghyZXNwb25zZRgBIAEoCzIqLmNlcmJvcy5yZXNwb25zZS52MS5DaGVja1Jlc291cmNlc1Jlc3BvbnNlQga6SAPIAQESOAoLYXVkaXRfdHJhaWwYAiABKAsyGy5jZXJib3MuYXVkaXQudjEuQXVkaXRUcmFpbEIGukgDyAEBIpYBChVQbGFuUmVzb3VyY2VzUmVzcG9uc2USQwoIcmVzcG9uc2UYASABKAsyKS5jZXJib3MucmVzcG9uc2UudjEuUGxhblJlc291cmNlc1Jlc3BvbnNlQga6SAPIAQESOAoLYXVkaXRfdHJhaWwYAiABKAsyGy5jZXJib3MuYXVkaXQudjEuQXVkaXRUcmFpbEIGukgDyAEBIqQBCgZCdW5kbGUSPwoIbWV0YWRhdGEYASABKAsyJS5jZXJib3MuY2xvdWQuZXBkcC52Mi5CdW5kbGUuTWV0YWRhdGFCBrpIA8gBARIQCghjb250ZW50cxgCIAEoDBpHCghNZXRhZGF0YRIbCglidW5kbGVfaWQYASABKAlCCLpIBXIDmAEQEh4KDXJ1bGVfcmV2aXNpb24YAiABKANCB7pIBCICIAAirQEKEEdldEJ1bmRsZVJlcXVlc3QSGQoHcnVsZV9pZBgBIAEoCUIIukgFcgOYAQwSIQoGc2NvcGVzGAIgAygJQhG6SA6SAQsQgAEYASIEcgIQARJFChFpZl9tb2RpZmllZF9zaW5jZRgDIAEoCzIlLmNlcmJvcy5jbG91ZC5lcGRwLnYyLkJ1bmRsZS5NZXRhZGF0YUgAiAEBQhQKEl9pZl9tb2RpZmllZF9zaW5jZSKEAQoRR2V0QnVuZGxlUmVzcG9uc2USLgoGYnVuZGxlGAEgASgLMhwuY2VyYm9zLmNsb3VkLmVwZHAudjIuQnVuZGxlSAASLgoMbm90X21vZGlmaWVkGAIgASgLMhYuZ29vZ2xlLnByb3RvYnVmLkVtcHR5SABCDwoGcmVzdWx0EgW6SAIIATKIAQoNQnVuZGxlU2VydmljZRJhCglHZXRCdW5kbGUSJi5jZXJib3MuY2xvdWQuZXBkcC52Mi5HZXRCdW5kbGVSZXF1ZXN0GicuY2VyYm9zLmNsb3VkLmVwZHAudjIuR2V0QnVuZGxlUmVzcG9uc2UiA5ACARoU+tLkkwIOEgxFWFBFUklNRU5UQUxCeAocZGV2LmNlcmJvcy5hcGkuY2xvdWQudjIuZXBkcFo9Z2l0aHViLmNvbS9jZXJib3MvY2xvdWQtYXBpL2dlbnBiL2NlcmJvcy9jbG91ZC9lcGRwL3YyO2VwZHB2MqoCGENlcmJvcy5BcGkuQ2xvdWQuVjIuRXBkcGIGcHJvdG8z",
+    "Ch9jZXJib3MvY2xvdWQvZXBkcC92Mi9lcGRwLnByb3RvEhRjZXJib3MuY2xvdWQuZXBkcC52MiKqBAoGQ29uZmlnEjkKCWV2YWx1YXRvchgBIAEoCzImLmNlcmJvcy5jbG91ZC5lcGRwLnYyLkNvbmZpZy5FdmFsdWF0b3ISMwoGc2NoZW1hGAIgASgLMiMuY2VyYm9zLmNsb3VkLmVwZHAudjIuQ29uZmlnLlNjaGVtYRruAQoJRXZhbHVhdG9yEkQKB2dsb2JhbHMYASADKAsyMy5jZXJib3MuY2xvdWQuZXBkcC52Mi5Db25maWcuRXZhbHVhdG9yLkdsb2JhbHNFbnRyeRIeChZkZWZhdWx0X3BvbGljeV92ZXJzaW9uGAIgASgJEhwKFGxlbmllbnRfc2NvcGVfc2VhcmNoGAMgASgIEhUKDWRlZmF1bHRfc2NvcGUYBCABKAkaRgoMR2xvYmFsc0VudHJ5EgsKA2tleRgBIAEoCRIlCgV2YWx1ZRgCIAEoCzIWLmdvb2dsZS5wcm90b2J1Zi5WYWx1ZToCOAEavgEKBlNjaGVtYRJECgtlbmZvcmNlbWVudBgBIAEoDjIvLmNlcmJvcy5jbG91ZC5lcGRwLnYyLkNvbmZpZy5TY2hlbWEuRW5mb3JjZW1lbnQibgoLRW5mb3JjZW1lbnQSGwoXRU5GT1JDRU1FTlRfVU5TUEVDSUZJRUQQABIUChBFTkZPUkNFTUVOVF9OT05FEAESFAoQRU5GT1JDRU1FTlRfV0FSThACEhYKEkVORk9SQ0VNRU5UX1JFSkVDVBADIkQKBUVycm9yEioKBGNvZGUYASABKA4yEC5nb29nbGUucnBjLkNvZGVCCrpIB4IBBBABIAASDwoHbWVzc2FnZRgCIAEoCSKoAQoITWV0YWRhdGESHwoOY2VyYm9zX3ZlcnNpb24YASABKAlCB7pIBHICEAESJAoSY2VyYm9zX2NvbW1pdF9oYXNoGAIgASgJQgi6SAVyA5gBKBIfCg13YXNtX2NoZWNrc3VtGAMgASgJQgi6SAVyA5gBQBI0CghidWlsdF9hdBgEIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXBCBrpIA8gBASIqChVMb2FkUnVsZVRhYmxlUmVzcG9uc2USEQoJYnVuZGxlX2lkGAEgASgJIpgBChZDaGVja1Jlc291cmNlc1Jlc3BvbnNlEkQKCHJlc3BvbnNlGAEgASgLMiouY2VyYm9zLnJlc3BvbnNlLnYxLkNoZWNrUmVzb3VyY2VzUmVzcG9uc2VCBrpIA8gBARI4CgthdWRpdF90cmFpbBgCIAEoCzIbLmNlcmJvcy5hdWRpdC52MS5BdWRpdFRyYWlsQga6SAPIAQEilgEKFVBsYW5SZXNvdXJjZXNSZXNwb25zZRJDCghyZXNwb25zZRgBIAEoCzIpLmNlcmJvcy5yZXNwb25zZS52MS5QbGFuUmVzb3VyY2VzUmVzcG9uc2VCBrpIA8gBARI4CgthdWRpdF90cmFpbBgCIAEoCzIbLmNlcmJvcy5hdWRpdC52MS5BdWRpdFRyYWlsQga6SAPIAQEipAEKBkJ1bmRsZRI/CghtZXRhZGF0YRgBIAEoCzIlLmNlcmJvcy5jbG91ZC5lcGRwLnYyLkJ1bmRsZS5NZXRhZGF0YUIGukgDyAEBEhAKCGNvbnRlbnRzGAIgASgMGkcKCE1ldGFkYXRhEhsKCWJ1bmRsZV9pZBgBIAEoCUIIukgFcgOYARASHgoNcnVsZV9yZXZpc2lvbhgCIAEoA0IHukgEIgIgACKtAQoQR2V0QnVuZGxlUmVxdWVzdBIZCgdydWxlX2lkGAEgASgJQgi6SAVyA5gBDBIhCgZzY29wZXMYAiADKAlCEbpIDpIBCxCAARgBIgRyAhABEkUKEWlmX21vZGlmaWVkX3NpbmNlGAMgASgLMiUuY2VyYm9zLmNsb3VkLmVwZHAudjIuQnVuZGxlLk1ldGFkYXRhSACIAQFCFAoSX2lmX21vZGlmaWVkX3NpbmNlIoQBChFHZXRCdW5kbGVSZXNwb25zZRIuCgZidW5kbGUYASABKAsyHC5jZXJib3MuY2xvdWQuZXBkcC52Mi5CdW5kbGVIABIuCgxub3RfbW9kaWZpZWQYAiABKAsyFi5nb29nbGUucHJvdG9idWYuRW1wdHlIAEIPCgZyZXN1bHQSBbpIAggBMogBCg1CdW5kbGVTZXJ2aWNlEmEKCUdldEJ1bmRsZRImLmNlcmJvcy5jbG91ZC5lcGRwLnYyLkdldEJ1bmRsZVJlcXVlc3QaJy5jZXJib3MuY2xvdWQuZXBkcC52Mi5HZXRCdW5kbGVSZXNwb25zZSIDkAIBGhT60uSTAg4SDEVYUEVSSU1FTlRBTEJ4ChxkZXYuY2VyYm9zLmFwaS5jbG91ZC52Mi5lcGRwWj1naXRodWIuY29tL2NlcmJvcy9jbG91ZC1hcGkvZ2VucGIvY2VyYm9zL2Nsb3VkL2VwZHAvdjI7ZXBkcHYyqgIYQ2VyYm9zLkFwaS5DbG91ZC5WMi5FcGRwYgZwcm90bzM",
     [
       file_buf_validate_validate,
       file_cerbos_audit_v1_audit,
@@ -372,6 +372,38 @@ export const MetadataSchema: GenMessage<
 > = /*@__PURE__*/ messageDesc(file_cerbos_cloud_epdp_v2_epdp, 2);
 
 /**
+ * @generated from message cerbos.cloud.epdp.v2.LoadRuleTableResponse
+ */
+export type LoadRuleTableResponse =
+  Message<"cerbos.cloud.epdp.v2.LoadRuleTableResponse"> & {
+    /**
+     * @generated from field: string bundle_id = 1;
+     */
+    bundleId: string;
+  };
+
+/**
+ * @generated from message cerbos.cloud.epdp.v2.LoadRuleTableResponse
+ */
+export type LoadRuleTableResponseJson = {
+  /**
+   * @generated from field: string bundle_id = 1;
+   */
+  bundleId?: string;
+};
+
+export type LoadRuleTableResponseValid = LoadRuleTableResponse;
+
+/**
+ * Describes the message cerbos.cloud.epdp.v2.LoadRuleTableResponse.
+ * Use `create(LoadRuleTableResponseSchema)` to create a new message.
+ */
+export const LoadRuleTableResponseSchema: GenMessage<
+  LoadRuleTableResponse,
+  { jsonType: LoadRuleTableResponseJson; validType: LoadRuleTableResponseValid }
+> = /*@__PURE__*/ messageDesc(file_cerbos_cloud_epdp_v2_epdp, 3);
+
+/**
  * @generated from message cerbos.cloud.epdp.v2.CheckResourcesResponse
  */
 export type CheckResourcesResponse =
@@ -428,7 +460,7 @@ export const CheckResourcesResponseSchema: GenMessage<
     jsonType: CheckResourcesResponseJson;
     validType: CheckResourcesResponseValid;
   }
-> = /*@__PURE__*/ messageDesc(file_cerbos_cloud_epdp_v2_epdp, 3);
+> = /*@__PURE__*/ messageDesc(file_cerbos_cloud_epdp_v2_epdp, 4);
 
 /**
  * @generated from message cerbos.cloud.epdp.v2.PlanResourcesResponse
@@ -484,7 +516,7 @@ export type PlanResourcesResponseValid =
 export const PlanResourcesResponseSchema: GenMessage<
   PlanResourcesResponse,
   { jsonType: PlanResourcesResponseJson; validType: PlanResourcesResponseValid }
-> = /*@__PURE__*/ messageDesc(file_cerbos_cloud_epdp_v2_epdp, 4);
+> = /*@__PURE__*/ messageDesc(file_cerbos_cloud_epdp_v2_epdp, 5);
 
 /**
  * @generated from message cerbos.cloud.epdp.v2.Bundle
@@ -538,7 +570,7 @@ export type BundleValid = Message<"cerbos.cloud.epdp.v2.Bundle"> & {
 export const BundleSchema: GenMessage<
   Bundle,
   { jsonType: BundleJson; validType: BundleValid }
-> = /*@__PURE__*/ messageDesc(file_cerbos_cloud_epdp_v2_epdp, 5);
+> = /*@__PURE__*/ messageDesc(file_cerbos_cloud_epdp_v2_epdp, 6);
 
 /**
  * @generated from message cerbos.cloud.epdp.v2.Bundle.Metadata
@@ -580,7 +612,7 @@ export type Bundle_MetadataValid = Bundle_Metadata;
 export const Bundle_MetadataSchema: GenMessage<
   Bundle_Metadata,
   { jsonType: Bundle_MetadataJson; validType: Bundle_MetadataValid }
-> = /*@__PURE__*/ messageDesc(file_cerbos_cloud_epdp_v2_epdp, 5, 0);
+> = /*@__PURE__*/ messageDesc(file_cerbos_cloud_epdp_v2_epdp, 6, 0);
 
 /**
  * @generated from message cerbos.cloud.epdp.v2.GetBundleRequest
@@ -632,7 +664,7 @@ export type GetBundleRequestValid = GetBundleRequest;
 export const GetBundleRequestSchema: GenMessage<
   GetBundleRequest,
   { jsonType: GetBundleRequestJson; validType: GetBundleRequestValid }
-> = /*@__PURE__*/ messageDesc(file_cerbos_cloud_epdp_v2_epdp, 6);
+> = /*@__PURE__*/ messageDesc(file_cerbos_cloud_epdp_v2_epdp, 7);
 
 /**
  * @generated from message cerbos.cloud.epdp.v2.GetBundleResponse
@@ -708,7 +740,7 @@ export type GetBundleResponseValid =
 export const GetBundleResponseSchema: GenMessage<
   GetBundleResponse,
   { jsonType: GetBundleResponseJson; validType: GetBundleResponseValid }
-> = /*@__PURE__*/ messageDesc(file_cerbos_cloud_epdp_v2_epdp, 7);
+> = /*@__PURE__*/ messageDesc(file_cerbos_cloud_epdp_v2_epdp, 8);
 
 /**
  * @generated from service cerbos.cloud.epdp.v2.BundleService